Added foreign key constraint trycatch to album deletion function, rearranged and added to design.

+++ Form1.cs
@@ -1,4 +1,7 @@
+using MySqlX.XDevAPI.Common;
 using System.Windows.Forms;
+using static System.ComponentModel.Design.ObjectSelectorEditor;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.TrackBar;
 
 namespace MusicSQLDatabaseApp
 {
@@ -51,7 +54,7 @@ namespace MusicSQLDatabaseApp
 
         private void label1_Click(object sender, EventArgs e)
         {
-
+            //do nothing
         }
 
         private void button3_Click(object sender, EventArgs e)
@@ -73,18 +76,58 @@ namespace MusicSQLDatabaseApp
         private void button4_Click(object sender, EventArgs e)
         {
             int rowClicked = dataGridView2.CurrentRow.Index;
-            MessageBox.Show("You have selected row" + rowClicked);
             int trackID = (int) dataGridView2.Rows[rowClicked].Cells[0].Value;
-            MessageBox.Show("You have selected track" + trackID);
 
             AlbumsDAO albumsDAO = new AlbumsDAO();
 
-            int result = albumsDAO.deleteTrack(trackID);
-            MessageBox.Show("You have deleted " + result);
+            DialogResult dialogResult = MessageBox.Show(
+                "You have selected track " + trackID + ".\n" +
+                "Are you sure you want to delete the current selection?",
+                "Delete Confirmation", MessageBoxButtons.YesNo);
 
-            dataGridView2.DataSource = null;
-            albums = albumsDAO.getAllAlbums();
+            if (dialogResult == DialogResult.Yes)
+            {
+                int result = albumsDAO.deleteTrack(trackID);
+                MessageBox.Show("You have deleted " + result + ".");
 
+                dataGridView2.DataSource = null;
+                albums = albumsDAO.getAllAlbums();
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+                MessageBox.Show("You have canceled deletion of track " + trackID + ".");
+            }
+        }
+
+        private void label9_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void button5_Click(object sender, EventArgs e)
+        {
+            int rowClicked = dataGridView1.CurrentRow.Index;
+            int albumID = (int)dataGridView1.Rows[rowClicked].Cells[0].Value;
+
+            AlbumsDAO albumsDAO = new AlbumsDAO();
+
+            DialogResult dialogResult = MessageBox.Show(
+                "You have selected album " + albumID + ".\n" +
+                "Are you sure you want to delete the current selection?",
+                "Delete Confirmation", MessageBoxButtons.YesNo);
+
+            if (dialogResult == DialogResult.Yes)
+            {
+                int result = albumsDAO.deleteAlbum(albumID);
+                MessageBox.Show("You have deleted " + result + ".");
+
+                dataGridView1.DataSource = null;
+                albums = albumsDAO.getAllAlbums();
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+                MessageBox.Show("You have canceled deletion of album " + albumID + ".");
+            }
         }
     }
 }
